leetcode

+++ .leetcode/9.回文数.cpp
@@ -16,17 +16,18 @@ public:
         if (x >= 0 && x < 10)
             return true;
 
-        stack<int> stack;
-        int digit = (int)log10(x) + 1;
+        stack<short int> stack;
+        short int digit = (short int)log10(x) + 1;
         int index(0);
 
         if ( digit%2 == 0 )
         {
+
             while(x != 0)
             {
                 if (index < digit/2)
                 {
-                    stack.push(x%10);
+                    stack.push((short int)(x%10));
                 }
                 else
                 {
@@ -47,7 +48,7 @@ public:
             {
                 if (index < digit/2)
                 {
-                    stack.push(x%10);
+                    stack.push((short int)(x%10));
                 }
                 else
                 {
